refactor: split Xref top-level types
