+++ src/CADShared/Cad/Database/Xrefs/XrefEx.cs
@@ -1,377 +1,4 @@
-// ReSharper disable ForCanBeConvertedToForeach
-
-#if AC_NET48 || ZWCAD
-using ArgumentNullException = Fs.Fox.Basal.ArgumentNullEx;
-#endif
-
 namespace Fs.Fox.Cad;
-
-#region 参照工厂
-
-/// <summary>
-/// 参照绑定模式接口
-/// </summary>
-public interface IXrefBindModes
-{
-    /// <summary>
-    /// 卸载
-    /// </summary>
-    public void Unload();
-
-    /// <summary>
-    /// 重载
-    /// </summary>
-    public void Reload();
-
-    /// <summary>
-    /// 拆离
-    /// </summary>
-    public void Detach();
-
-    /// <summary>
-    /// 绑定
-    /// </summary>
-    public void Bind();
-}
-
-/// <summary>
-/// 参照工厂类
-/// </summary>
-/// <param name="tr"></param>
-/// <param name="xrefNames">要处理的参照名称,<see langword="null"/>就处理所有</param>
-public class XrefFactory(DBTrans tr, HashSet<string>? xrefNames = null) : IXrefBindModes
-{
-    #region 私有字段
-
-    private readonly DBTrans _tr = tr;
-
-    /// <summary>
-    /// 要处理的参照名称,<see langword="null"/>就处理所有
-    /// </summary>
-    private readonly HashSet<string>? _xrefNames = xrefNames;
-
-    #endregion
-
-    #region 公开字段
-
-    /// <summary>
-    /// 解析外部参照:线性引擎<br/>
-    /// 默认<see langword="false"/><br/>
-    /// <see langword="true"/>时会在cad命令历史打印一些AEC信息,并导致绑定慢一点...具体作用不详<br/>
-    /// </summary>
-    public readonly bool UseThreadEngine = false;
-
-    /// <summary>
-    /// 解析外部参照:仅处理 Unresolved_未融入(未解析)的参照<br/>
-    /// 默认<see langword="true"/>
-    /// </summary>
-    public readonly bool DoNewOnly = true;
-
-    /// <summary>
-    /// 解析外部参照:包含僵尸参照
-    /// </summary>
-    public readonly bool IncludeGhosts = true;
-
-
-    /// <summary>
-    /// 绑定模式和双美元符号相关(与cad保持相同地默认)<br/>
-    /// <see langword="false"/>为绑定模式,产生双美元;
-    /// <see langword="true"/>为插入模式,块重名会以本图覆盖;
-    /// </summary>
-    public readonly bool BindOrInsert = false;
-
-    /// <summary>
-    /// bind时候是否拆离参照<br/>
-    /// 默认<see langword="true"/>:学官方的绑定后自动拆离
-    /// </summary>
-    public readonly bool AutoDetach = true;
-
-    /// <summary>
-    /// bind时候是否删除被卸载的嵌套参照<br/>
-    /// 默认<see langword="true"/>
-    /// </summary>
-    public readonly bool EraseNested = true;
-
-    /// <summary>
-    /// bind时候控制绑定的符号表:请保持默认<br/>
-    /// 目前仅推荐用于<see cref="SymModes.LayerTable"/>项<br/>
-    /// 其他项有异常:<see langword="eWasOpenForNotify"/><br/>
-    /// </summary>
-    public readonly SymModes SymModesBind = SymModes.LayerTable;
-
-    #endregion
-
-
-    #region 重写
-
-    /// <summary>
-    /// 绑定
-    /// </summary>
-    public void Bind()
-    {
-        // 此功能有绑定出错的问题
-        // db.BindXrefs(xrefIds, true);
-
-        // 绑定后会自动拆离
-        // 此功能修补了上面缺失
-        DoubleBind();
-    }
-
-    /// <summary>
-    /// 分离
-    /// </summary>
-    public void Detach()
-    {
-        using ObjectIdCollection xrefIds = new();
-        GetAllXrefNode(xrefIds);
-        foreach (ObjectId id in xrefIds)
-            _tr.Database.DetachXref(id);
-    }
-
-    /// <summary>
-    /// 重载
-    /// </summary>
-    public void Reload()
-    {
-        using ObjectIdCollection xrefIds = new();
-        GetAllXrefNode(xrefIds);
-        if (xrefIds.Count > 0)
-            _tr.Database.ReloadXrefs(xrefIds);
-    }
-
-    /// <summary>
-    /// 卸载
-    /// </summary>
-    public void Unload()
-    {
-        using ObjectIdCollection xrefIds = new();
-        GetAllXrefNode(xrefIds);
-        if (xrefIds.Count > 0)
-            _tr.Database.UnloadXrefs(xrefIds);
-    }
-
-    #endregion
-
-    #region 双重绑定
-
-    /// <summary>
-    /// 获取参照
-    /// </summary>
-    /// <param name="xrefIds">返回全部参照id</param>
-    private void GetAllXrefNode(ObjectIdCollection xrefIds)
-    {
-        // 储存要处理的参照id
-        //var xrefIds = new ObjectIdCollection();
-        XrefNodeForEach((xNodeName, xNodeId, _, _) =>
-        {
-            if (XrefNamesContains(xNodeName))
-                xrefIds.Add(xNodeId);
-        });
-    }
-
-    private bool XrefNamesContains(string xNodeName)
-    {
-        // 为空的时候全部加入 || 有内容时候含有目标
-        return _xrefNames is null || _xrefNames.Contains(xNodeName);
-    }
-
-    /// <summary>
-    /// 遍历参照
-    /// </summary>
-    /// <param name="action">(参照名,参照块表记录id,参照状态,是否嵌入)</param>
-    private void XrefNodeForEach(Action<string, ObjectId, XrefStatus, bool> action)
-    {
-        // btRec.IsFromOverlayReference 是覆盖
-        // btRec.GetXrefDatabase(true) 外部参照数据库
-
-
-        // 解析外部参照:此功能不能锁定文档
-        _tr.Database.ResolveXrefs(UseThreadEngine, DoNewOnly);
-
-        var xg = _tr.Database.GetHostDwgXrefGraph(IncludeGhosts);
-        for (var i = 0; i < xg.NumNodes; i++)
-        {
-            var xNode = xg.GetXrefNode(i);
-            if (!xNode.BlockTableRecordId.IsOk())
-                continue;
-
-            action.Invoke(xNode.Name, xNode.BlockTableRecordId, xNode.XrefStatus, xNode.IsNested);
-        }
-    }
-
-    /// <summary>
-    /// 符号表记录加入容器
-    /// </summary>
-    private static void AddedXBindIds<TTable, TRecord>(ObjectIdCollection xbindXrefsIds,
-        SymbolTable<TTable, TRecord> symbolTable) where TTable : SymbolTable
-        where TRecord : SymbolTableRecord, new()
-    {
-        symbolTable.ForEach(tabRec =>
-        {
-            if (tabRec.IsResolved)
-                xbindXrefsIds.Add(tabRec.ObjectId);
-        }, checkIdOk: true);
-    }
-
-
-    private void GetXBindIds(ObjectIdCollection xbindIds)
-    {
-        // xbind
-        // 0x01 它是用来绑其他符号表,绑块表会有异常
-        // 0x02 集合若有问题,就会出现eWrongObjectType
-        //var xbindIds = new ObjectIdCollection();
-
-        // 起初测试是将九大符号表记录均加入的,但经实测不行...(为什么?存疑)
-
-        #region Option1
-
-        if ((SymModesBind & SymModes.LayerTable) == SymModes.LayerTable)
-            AddedXBindIds(xbindIds, _tr.LayerTable);
-
-        if ((SymModesBind & SymModes.TextStyleTable) == SymModes.TextStyleTable)
-            AddedXBindIds(xbindIds, _tr.TextStyleTable);
-
-        if ((SymModesBind & SymModes.RegAppTable) == SymModes.RegAppTable)
-            AddedXBindIds(xbindIds, _tr.RegAppTable);
-
-        if ((SymModesBind & SymModes.DimStyleTable) == SymModes.DimStyleTable)
-            AddedXBindIds(xbindIds, _tr.DimStyleTable);
-
-        if ((SymModesBind & SymModes.LinetypeTable) == SymModes.LinetypeTable)
-            AddedXBindIds(xbindIds, _tr.LinetypeTable);
-
-        #endregion
-
-        #region Option2
-
-        if ((SymModesBind & SymModes.UcsTable) == SymModes.UcsTable)
-            AddedXBindIds(xbindIds, _tr.UcsTable);
-
-        if ((SymModesBind & SymModes.ViewTable) == SymModes.ViewTable)
-            AddedXBindIds(xbindIds, _tr.ViewTable);
-
-        if ((SymModesBind & SymModes.ViewportTable) == SymModes.ViewportTable)
-            AddedXBindIds(xbindIds, _tr.ViewportTable);
-
-        #endregion
-    }
-
-    private void GetBindIds(ObjectIdCollection bindIds)
-    {
-        // bind 只绑块表
-        //var bindIds = new ObjectIdCollection();
-
-        _tr.BlockTable.ForEach(btr =>
-        {
-            if (btr.IsLayout)
-                return;
-
-            // 外部参照 && 已融入
-            if (btr is { IsFromExternalReference: true, IsResolved: true })
-                bindIds.Add(btr.ObjectId);
-        }, checkIdOk: true);
-    }
-
-    /// <summary>
-    /// 获取可以拆离的ids
-    /// </summary>
-    /// <param name="nested">返回已卸载中含有嵌套的参照,要重载之后才能绑定</param>
-    /// <returns>返回未参照中嵌套的参照,直接拆离</returns>
-    private List<ObjectId> GetDetachIds(Dictionary<ObjectId, string> nested)
-    {
-        // 直接拆离的id
-        List<ObjectId> detachIds = [];
-
-        // 收集要处理的id
-        XrefNodeForEach((xNodeName, xNodeId, xNodeStatus, xNodeIsNested) =>
-        {
-            switch (xNodeStatus)
-            {
-                case XrefStatus.Unresolved: // 未融入_ResolveXrefs参数2
-                    break;
-                case XrefStatus.FileNotFound: // 未融入(未解析)_未找到文件
-                    break;
-                case XrefStatus.Unreferenced: // 未参照
-                {
-                    // 为空的时候全部加入 || 有内容时候含有目标
-                    if (XrefNamesContains(xNodeName))
-                        detachIds.Add(xNodeId);
-                }
-                    break;
-                case XrefStatus.Unloaded: // 已卸载
-                {
-                    // 为空的时候全部加入 || 有内容时候含有目标
-                    if (XrefNamesContains(xNodeName))
-                    {
-                        var btr = _tr.GetObject<BlockTableRecord>(xNodeId);
-                        if (btr != null && btr.IsFromExternalReference)
-                        {
-                            if (!xNodeIsNested)
-                                detachIds.Add(xNodeId);
-                            else if (!nested.ContainsKey(xNodeId))
-                                nested.Add(xNodeId, xNodeName); // 嵌套参照
-                        }
-                    }
-                }
-                    break;
-                case XrefStatus.Resolved: // 已融入_就是可以绑定的
-                    break;
-                case XrefStatus.NotAnXref: // 不是外部参照
-                    break;
-            }
-        });
-        return detachIds;
-    }
-
-    /// <summary>
-    /// 双重绑定参照
-    /// <a href="https://www.cnblogs.com/SHUN-ONCET/p/16593360.html">参考链接</a>
-    /// </summary>
-    private void DoubleBind()
-    {
-        Dictionary<ObjectId, string> nested = new();
-        var detachIds = GetDetachIds(nested);
-
-        // 拆离:未参照的文件
-        if (AutoDetach)
-        {
-            for (var i = 0; i < detachIds.Count; i++)
-                _tr.Database.DetachXref(detachIds[i]);
-        }
-
-        // 重载:嵌套参照已卸载了,需要重载之后才能进行绑定
-        var keys = nested.Keys;
-        if (keys.Count > 0)
-        {
-            using ObjectIdCollection idc = new(keys.ToArray());
-            _tr.Database.ReloadXrefs(idc);
-        }
-
-        // 绑定:切勿交换,否则会绑定无效
-        using ObjectIdCollection bindIds = new();
-        using ObjectIdCollection xbindIds = new();
-        GetBindIds(bindIds);
-        GetXBindIds(xbindIds);
-        if (xbindIds.Count > 0)
-            _tr.Database.XBindXrefs(xbindIds, BindOrInsert);
-        if (bindIds.Count > 0)
-            _tr.Database.BindXrefs(bindIds, BindOrInsert);
-
-
-        // 内部删除嵌套参照的块操作
-        if (EraseNested)
-        {
-            foreach (var item in nested)
-            {
-                var name = item.Value;
-                if (_tr.BlockTable.Has(name))
-                    _tr.GetObject<BlockTableRecord>(_tr.BlockTable[name], OpenMode.ForWrite)?.Erase();
-            }
-        }
-    }
-
-    #endregion
-}
 
 /// <summary>
 /// 参照扩展
@@ -407,278 +34,3 @@ public static class XrefEx
         });
     }
 }
-
-#endregion
-
-#region 参照路径工具类
-
-/// <summary>
-/// 获取外部参照的路径
-/// </summary>
-public class XrefPath
-{
-    #region 属性
-
-    /// <summary>
-    /// 基础路径
-    /// </summary>
-    public readonly string? CurrentDatabasePath;
-
-    /// <summary>
-    /// 是否外部参照
-    /// </summary>
-    public bool IsFromExternalReference { get; private set; }
-
-    /// <summary>
-    /// 外部参照保存的路径
-    /// <para>
-    /// 它们会是以下任一路径:<br/>
-    /// 0x01 相对路径<br/>
-    /// 0x02 绝对路径<br/>
-    /// 0x03 共目录优先找到的路径(文件夹整体移动会发生此类情况)
-    /// </para>
-    /// </summary>
-    public string? PathSave { get; private set; }
-
-    /// <summary>
-    /// 找到的路径(参照面板的名称)
-    /// <para><see cref="PathSave"/>路径不存在时,返回是外部参照dwg文件路径</para>
-    /// </summary>
-    public string? PathDescribe { get; private set; }
-
-    private string? _pathComplete;
-
-    /// <summary>
-    /// 绝对路径
-    /// </summary>
-    public string? PathComplete =>
-        _pathComplete ??= PathConverter(CurrentDatabasePath, PathDescribe, PathConverterModes.Complete);
-
-    private string? _pathRelative;
-
-    /// <summary>
-    /// 相对路径
-    /// </summary>
-    public string? PathRelative =>
-        _pathRelative ??= PathConverter(CurrentDatabasePath, PathComplete, PathConverterModes.Relative);
-
-    #endregion
-
-    #region 构造
-
-    /// <summary>
-    /// 获取外部参照的路径
-    /// </summary>
-    /// <param name="brf">外部参照图元</param>
-    /// <param name="tr">事务</param>
-    /// <returns>是否外部参照</returns>
-    public XrefPath(BlockReference brf, DBTrans tr)
-    {
-        //if (brf == null)
-        //    throw new ArgumentNullException(nameof(brf));
-        ArgumentNullException.ThrowIfNull(brf);
-        CurrentDatabasePath = Path.GetDirectoryName(tr.Database.Filename);
-
-        var btRec = tr.GetObject<BlockTableRecord>(brf.BlockTableRecord); // 块表记录
-        if (btRec == null)
-            return;
-
-        IsFromExternalReference = btRec.IsFromExternalReference;
-        if (!IsFromExternalReference)
-            return;
-
-        // 相对路径==".\\AA.dwg"
-        // 无路径=="AA.dwg"
-        PathSave = btRec.PathName;
-
-        if ((!string.IsNullOrEmpty(PathSave) && PathSave[0] == '.') || File.Exists(PathSave))
-        {
-            // 相对路径||绝对路径
-            PathDescribe = PathSave;
-        }
-        else
-        {
-            // 无路径
-            var db = btRec.GetXrefDatabase(true);
-            PathDescribe = db.Filename;
-        }
-    }
-
-    #endregion
-
-    #region 静态函数
-
-    /// <summary>
-    /// 获取相对路径或者绝对路径
-    /// <see href="https://www.cnblogs.com/hont/p/5412340.html">参考链接</see>
-    /// </summary>
-    /// <param name="directory">基础目录(末尾无斜杠)</param>
-    /// <param name="fileRelations">相对路径或者绝对路径</param>
-    /// <param name="converterModes">依照枚举返回对应的字符串</param>
-    /// <returns></returns>
-    public static string? PathConverter(string? directory, string? fileRelations,
-        PathConverterModes converterModes)
-    {
-        //if (directory == null)
-        //    throw new ArgumentNullException(nameof(directory));
-        //if (fileRelations == null)
-        //    throw new ArgumentNullException(nameof(fileRelations));
-
-        ArgumentNullException.ThrowIfNull(directory);
-        ArgumentNullException.ThrowIfNull(fileRelations);
-
-        string? result = null;
-        switch (converterModes)
-        {
-            case PathConverterModes.Relative:
-                result = GetRelativePath(directory, fileRelations);
-                break;
-            case PathConverterModes.Complete:
-                result = GetCompletePath(directory, fileRelations);
-                break;
-        }
-
-        return result;
-    }
-
-#if error_demo
-    /// <summary>
-    /// 绝对路径->相对路径
-    /// </summary>
-    /// <param name="strDbPath">绝对路径</param>
-    /// <param name="strXrefPath">相对关系</param>
-    /// <returns></returns>
-    /// StringHelper.GetRelativePath("G:\\A1.项目\\20190920金山谷黄宅\\01.饰施图\\03.平面图",
-    /// "G:\\A1.项目\\20190920金山谷黄宅\\01.饰施图\\01.辅助文件\\图框\\A3图框.dwg");
-    public static string GetRelativePath(string strDbPath, string strXrefPath)
-    {
-        Uri uri1 = new(strXrefPath);
-        Uri uri2 = new(strDbPath);
-        Uri relativeUri = uri2.MakeRelativeUri(uri1);
-        // 测试例子变成 01.%E8%BE%85%E5%8A%A9%E6%96%87%E4%BB%B6/%E5%9B%BE%E6%A1%86/A3%E5%9B%BE%E6%A1%86.dwg
-        string str = relativeUri.ToString();
-
-        // 因为这里不会实现".\A.dwg"而是"A.dwg",所以加入这个操作,满足同目录文件
-        var strs = str.Split('\\');
-        if (strs.Length == 1)
-            str = ".\\" + str;
-        return str;
-    }
-#else
-    /// <summary>
-    /// 绝对路径->相对路径
-    /// </summary>
-    /// <param name="directory">相对关系:文件夹路径</param>
-    /// <param name="file">完整路径:文件路径</param>
-    /// <returns>相对路径</returns>
-    /// <![CDATA[
-    /// GetRelativePath("G:\\A1.项目\\20190920金山谷黄宅\\01.饰施图\\03.平面图",
-    /// "G:\\A1.项目\\20190920金山谷黄宅\\01.饰施图\\01.辅助文件\\图框\\A3图框.dwg")
-    /// =>  "..\\01.辅助文件\\图框\\A3图框.dwg"
-    /// ]]>
-    private static string GetRelativePath(string directory, string file)
-    {
-        string[] directories = directory.Split('\\');
-        string[] files = file.Split('\\');
-        // 获取两条路径中的最短路径
-        var getMinLength = directories.Length < files.Length ? directories.Length : files.Length;
-
-        // 用于确定我们退出的循环中的位置。
-        var lastCommonRoot = -1;
-        int index;
-        // 找到共根
-        for (index = 0; index < getMinLength; index++)
-        {
-            if (directories[index] != files[index])
-                break;
-            lastCommonRoot = index;
-        }
-
-        // 如果我们没有找到一个共同的前缀,那么抛出
-        if (lastCommonRoot == -1)
-            throw new ArgumentException("路径没有公共相同路径部分");
-
-        // 建立相对路径
-        var result = new StringBuilder();
-        for (index = lastCommonRoot + 1; index < directories.Length; index++)
-            if (directories[index].Length > 0)
-                result.Append("..\\"); // 上级目录加入
-
-        // 添加文件夹
-        for (index = lastCommonRoot + 1; index < files.Length - 1; index++)
-            result.Append(files[index] + "\\");
-
-        // 本级目录
-        if (result.Length == 0)
-            result.Append(".\\");
-        // result.Append(strXrefPaths[^1]);// 下级目录加入
-        result.Append(files[^1]); // 下级目录加入
-        return result.ToString();
-    }
-#endif
-
-    /// <summary>
-    /// 相对路径->绝对路径
-    /// </summary>
-    /// <param name="directory">文件夹路径</param>
-    /// <param name="relativePath">相对关系:有..的</param>
-    /// <returns>完整路径</returns>
-    /// <![CDATA[
-    /// GetCompletePath("G:\\A1.项目\\20190920金山谷黄宅\\01.饰施图\\03.平面图" ,
-    /// "..\\01.辅助文件\\图框\\A3图框.dwg")
-    /// =>   "G:\\A1.项目\\20190920金山谷黄宅\\01.饰施图\\01.辅助文件\\图框\\A3图框.dwg"
-    /// ]]>
-    private static string? GetCompletePath(string directory, string relativePath)
-    {
-        if (relativePath.Trim() == string.Empty)
-            return null;
-
-        var relativeName = Path.GetDirectoryName(relativePath);
-        if (relativeName is null)
-            return null;
-
-        if (relativePath[0] != '.')
-            return relativePath;
-
-        const char slash = '\\';
-
-        // 判断向上删除几个
-        var slashes = relativeName.Split(slash);
-        var index = 0;
-        // ReSharper disable once ForCanBeConvertedToForeach
-        for (var i = 0; i < slashes.Length; i++)
-        {
-            if (slashes[i] != "..")
-                break;
-            index++;
-        }
-
-        var result = new StringBuilder();
-        // 前段
-        var pathDwgs = directory.Split(slash);
-        pathDwgs = pathDwgs.Where(s => !string.IsNullOrEmpty(s)).ToArray(); // 清理空数组
-        for (var i = 0; i < pathDwgs.Length - index; i++)
-        {
-            result.Append(pathDwgs[i]);
-            result.Append(slash);
-        }
-
-        // 后段
-        for (var i = 0; i < slashes.Length; i++)
-        {
-            var item = slashes[i];
-            if (item != "." && item != "..")
-            {
-                result.Append(item);
-                result.Append(slash);
-            }
-        }
-
-        result.Append(Path.GetFileName(relativePath));
-        return result.ToString();
-    }
-
-    #endregion
-}
-
-#endregion
